Updated comments, deleted old code
I updated some comments surrounding the ratios and chi-square tests that we need to review the code for reproducibility purposes. I also deleted some of the code from when we tried to make bar graphs with confidence intervals on them to make the do-file more accessible, but I did store it in the google doc with the unemployment rate inputs.